# dist/api.yml
...
openapi: 3.0.0
info:
  title: Soccer API
  description: This is our soccer API project
  contact:
    email: naorsu@post.bgu.ac.il
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
paths:
  /auth/register:
    post:
      tags:
      - auth
      summary: add a new user
      description: A json object that contains all user info is added to the data
      requestBody:
        description: User info to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
        required: true
      responses:
        "201":
          description: User has been created
        "400":
          description: invalid input
        "409":
          description: User is already exists
  /auth/login:
    post:
      tags:
      - auth
      summary: login to server
      description: sends user name and password to verify to user cardntials
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body'
        required: true
      responses:
        "200":
          description: |
            User was found in the DB and authenitcated successfully. The session ID is returned in a cookie named `session`.  You should include this cookie in subsequent requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "400":
          description: invalid input
        "404":
          description: couldnt find user name matching that password
  /auth/logout:
    post:
      tags:
      - auth
      summary: logout from server
      description: deltes user cookies and authntication
      responses:
        "200":
          description: |
            User was loggedout successfully. The session ID is rested in a cookie named `session`.
        "400":
          description: no user is logged in
...
  /league/summary:
    get:
      tags:
      - league
      summary: get league current summrized info
      description: returns an json object with current league params
      responses:
        "200":
          description: returns an json object with current league params
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
        "400":
          description: didnt find avilable data on league
...
  /league/current_fixture:
    get:
      tags:
      - league
      summary: get league current fixture
      description: returns a json object with two arrays- array of pre played matches of the current fixture  and a array of post played matches of the current fixture
      responses:
        "200":
          description: returns a json object with two arrays- array of pre played matches of the current fixture  and a array of post played matches of the current fixture
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_2'
        "400":
          description: didnt find avilable data on league
  /league/matches/{league_id}:
      get:
        tags:
        - league
        summary: get league season matches
        description: returns a json object with array of all current season matches
        parameters:
          - name: league_id
            in: path
            required: true
            style: simple
            explode: false
            schema:
              type: string
        responses:
          "200":
            description: returns a json object with array of all current season matches
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/inline_response_200_3'
          "400":
            description: didnt find avilable data on league
          "404":
            description: couldn't find matches in the given league_id
...
components:
  schemas:
    user:
      required:
      - country
      - email
      - firstname
      - lastname
      - password
      - profile_picture
      - username
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: Must have 3-8 characters, letters only
          example: KingMessi
        firstname:
          type: string
          example: May
        lastname:
          type: string
          example: Suban
        country:
          type: string
        password:
          maxLength: 10
          minLength: 5
          type: string
          description: Must have 5-10 characters, should contain at least one digit and one special character
          format: password
          example: my*pass21
        email:
          type: string
          format: email
          example: Suban@gmail.com
        profile picture:
          type: string
          description: Link to the profile image, recommended to add picture from cloudinary
          format: URL
          example: https://res.cloudinary.com/dswkzsdoq/image/upload/v1620738869/20171122_132608_zvrw97.jpg
    event:
      required:
      - date
      - details
      - event_type
      - min_in_game
      type: object
      properties:
        date:
          type: string
          description: The event's date
          format: date
        min_in_game:
          type: string
          description: The minute in the game when the event occured
          example: "12"
        event_type:
          type: string
          description: The type of the event
          example: yellow card
        details:
          type: string
          description: The description of the event
          example: Messi foueled Ronaldo and got a yellow card
    match:
      required:
      - away_team
      - court
      - date
      - home_team
      - match_id
      - referee_name
      type: object
      properties:
        match_id:
          type: string
          description: unique match id
          example: RX23231F
        home_team:
          type: string
          description: The host team's name
          example: Hapoel Tel-Aviv
        away_team:
          type: string
          description: The guest team's name
          example: Hapoel Beer-Sheva
        court:
          type: string
          description: The court's id
          example: "1121"
        season:
          type: string
          description: The season
          example: "2020/2021"
        referee_name:
          type: string
          description: The referee's name
          example: Ron
        date:
          type: string
          description: The match's date
          format: date
        eventlog:
          type: array
          description: The log events of the match
          items:
            $ref: '#/components/schemas/event'
        score:
          type: string
          description: The match's score, (home_team-away_team)
          example: 3-0
...
    body:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: Must have 3-8 characters, letters only
          example: KingMessi
        password:
          maxLength: 10
          minLength: 5
          type: string
          description: Must have 5-10 characters, should contain at least one digit and one special character
          format: password
          example: my*pass21
    inline_response_200:
      type: object
      properties:
        roles:
          type: array
          description: The role list of the user
          example:
          - referee
          - union represntative
          - player
          items:
            type: string
...
    inline_response_200_1:
      type: object
      properties:
        season_name:
          type: string
          description: current runnig season
          example: 2012/2013
        stage_name:
          type: string
          description: current runing fixture
          example: fixture 14
        match:
          $ref: '#/components/schemas/match'
...
    inline_response_200_2:
      type: object
      properties:
        pre_played_matches:
          type: array
          description: pre played matches
          items:
            $ref: '#/components/schemas/match'
        post_played_matches:
          type: array
          description: post played matches
          items:
            $ref: '#/components/schemas/match'
    inline_response_200_3:
      type: object
      properties:
        season_matches:
          type: array
          description: all season matches
          items:
            $ref: '#/components/schemas/match'
...